Add a simple Flask app for REST search
We need JSON results for search queries to serve up via AngularJS or
whatever else we decide to use to build the front end. For now, go with
a simple plainto_tsquery() call under the covers; it can handle almost
any input, at the expense of not supporting more advanced features like
booleans and wildcards.

Signed-off-by: Dan Scott <dscott@laurentian.ca>