Add in year and collection facets
The collection facets might not turn out to be good enough, as it
provides no distinction between Reaction and L'Orignal. This might lead
to having to provide more granularity in our data store (perhaps
creating a new column based on the titles?). In any case, we now have
year and collection facets available to be added to the display.

Signed-off-by: Dan Scott <dscott@laurentian.ca>